Cleansing unused files and update docker-compose

+++ app-laravel/database/migrations/2021_01_26_145731_create_role_users_table.php
@@ -14,8 +14,8 @@ class CreateRoleUsersTable extends Migration
     public function up()
     {
         Schema::create('role_users', function (Blueprint $table) {
-            $table->unsignedInteger('user_id');
-            $table->unsignedInteger('role_id');
+            $table->unsignedBigInteger('user_id');
+            $table->unsignedBigInteger('role_id');
             $table->timestamps();
 
             $table->unique(['user_id', 'role_id']);
